Added buttifier from dentist/buttebot.
still needs a setting to actually work. Also I don't want to steal files form
buttebot, so I'm putting in some pull request for a setup.py in buttebot so the
modules can be reused. For now, put the files "buttifier.py", "hypenate.py" and
"prob.py" from buttbot in modules/buttes/ to have it work.